split BuildingHours into the fetching and the listing components
